+++ assets/js/main.js
@@ -1,18 +1,18 @@
-import { retrieveUserData, appendUserDataToTable } from './utils.js';
-import { saveUserDataToLocal, loadUserDataFromLocal } from './localStorage.js';
+//import { retrieveUserData, appendUserDataToTable } from './utils.js';
+    //import { saveUserDataToLocal, loadUserDataFromLocal } from './localStorage.js';
 
 // Main function to handle user data
-function handleUserData() {
-    let userData = retrieveUserData();
-    appendUserDataToTable(userData);
-    saveUserDataToLocal(userData);
-}
+//function handleUserData() {
+    //let userData = retrieveUserData();
+    //appendUserDataToTable(userData);
+    //saveUserDataToLocal(userData);
+//}
 
 
-document.getElementById('submitUserData').addEventListener('click', handleUserData);
+//document.getElementById('submitUserData').addEventListener('click', handleUserData);
 
 // Load and display existing data from localStorage on page load
-document.addEventListener('DOMContentLoaded', () => {
-    const storedData = loadUserDataFromLocal();
-    storedData.forEach(userData => appendUserDataToTable(userData));
-});
+//document.addEventListener('DOMContentLoaded', () => {
+   // const storedData = loadUserDataFromLocal();
+   // storedData.forEach(userData => appendUserDataToTable(userData));
+//});
