+++ index.html
@@ -17,35 +17,45 @@
             <form id="zodiac-form" class="text-center">
                 <div class="row">
                     <div class="col-sm-12 col-md-6">
-                        <label for="name">Name:</label>
+                        <label for="userName">Name:</label>
                     </div>
                 </div>
                 <div class="row">
                     <div class="col-sm-12 col-md-6">
-                        <input type="text" id="name" name="user-name">
+                        <input type="text" id="userName" name="user-name">
                     </div>
                 </div>
                 <div class="row">
                     <div class="col-sm-12 col-md-6">
-                        <label for="email">Zodiac Sign:</label>
+                        <label for="userZodiac">Zodiac Sign:</label>
                     </div>
                 </div>
                 <div class="row">
-                    <div class="col-sm-12 col-md-6"><input type="text" id="sign" name="user-sign"></div>
+                    <div class="col-sm-12 col-md-6"><input type="text" id="userZodiac" name="user-sign"></div>
                 </div>
                 <button>Submit</button>
             </form>
+        </section>
+        <button id="goToCards" class="btn-primary">
+            View Sign Info
+
+        </button>
+        <section id="zodiacCardsContainer">
+            
         </section>
     </main>
 
 
     <script src="./assets/js/index.js"></script>
+    <script src="./assets/js/local.js"></script>
+    <script src="./assets/js/utils.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.8/dist/umd/popper.min.js"
         integrity="sha384-I7E8VVD/ismYTF4hNIPjVp/Zjvgyol6VFvRkX/vR+Vc4jQkC+hVqc2pM8ODewa9r"
         crossorigin="anonymous"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.min.js"
         integrity="sha384-0pUGZvbkm6XF6gxjEnlmuGrJXVbNuzT9qBBavbLwCsOGabYfZo0T0to5eqruptLy"
         crossorigin="anonymous"></script>
+
 </body>
